refactor(backend): change file names
