+++ packages/backend/src/open_payments/incoming/routes.test.ts
@@ -12,7 +12,7 @@ import { resetGraphileDb } from '../../tests/graphileDb'
 import { GraphileProducer } from '../../messaging/graphileProducer'
 import { Config, IAppConfig } from '../../config/app'
 import { IocContract } from '@adonisjs/fold'
-import { initIocContainer } from '../../'
+import { initIocContainer } from '../..'
 import { AppServices } from '../../app'
 import { truncateTables } from '../../tests/tableManager'
 import { randomAsset } from '../../tests/asset'
