+++ packages/backend/src/open_payments/incoming/service.test.ts
@@ -15,7 +15,7 @@ import { resetGraphileDb } from '../../tests/graphileDb'
 import { GraphileProducer } from '../../messaging/graphileProducer'
 import { Config } from '../../config/app'
 import { IocContract } from '@adonisjs/fold'
-import { initIocContainer } from '../../'
+import { initIocContainer } from '../..'
 import { AppServices } from '../../app'
 import { Pagination } from '../../shared/baseModel'
 import { getPageTests } from '../../shared/baseModel.test'
